Implement layout with top bar and (future) sidebar
