+++ src/assets/style.scss
@@ -11,14 +11,24 @@ $fg-strong: #eee;
 $fg-link: #4fbdff;
 $fg-link-visited: #ba9bdf;
 $border-color: #39434d;
+$border-color-header: #184851;
 $border-radius: 2px;
 $padding-containers: 20px;
 $padding-containers-small: 10px;
-$font-family: "Helvetica", sans-serif;
 $font-size: 12pt;
+$font-size-logo: 14pt;
+$font-family: "Helvetica", sans-serif;
+$font-family-headers: "ConvectionRegular", "Helvetica", sans-serif;
 $font-family-code: "Fira Code", monospace;
 $font-size-code: 10pt;
-$width-content: 65em;
+$width-content: 60em;
+$transition-time: 0.3s;
+
+@font-face {
+  font-family: ConvectionRegular;
+  src: local("Convection"), url("/assets/ConvectionRegular.ttf");
+  font-weight: normal;
+}
 
 body {
   color: $fg;
@@ -72,11 +82,22 @@ h4 {font-size: 1.00em;}
 h5 {font-size: 0.83em;}
 h6 {font-size: 0.75em;}
 
-main {
-  max-width: $width-content;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
+input {
+  border-radius: $border-radius;
+  border: 1px solid $border-color;
+  background: $bg-inset;
+  padding: $padding-containers-small;
+  font-size: $font-size;
+  color: $fg;
+
+  &:disabled {
+    cursor: not-allowed;
+    background: $bg-content;
+    transition: background $transition-time;
+    &:hover {
+      background: $bg-overlay;
+    }
+  }
 }
 
 hr {
@@ -89,19 +110,64 @@ details summary {
 }
 
 h1.page-title {
-  font-size: 2em;
+  font-size: 2.5em;
 }
 
 a.header-anchor {
   color: $fg-faded;
   opacity: 0;
   padding: 0 0.5em;
-  transition: opacity 0.3s;
+  transition: opacity $transition-time;
 }
 
 h1, h2, h3, h4, h5, h6 {
+  font-family: $font-family-headers;
   &:hover a.header-anchor {
     opacity: 1;
+  }
+}
+
+h1 {
+  border-bottom: 1px solid $border-color;
+  padding-bottom: 10px;
+  overflow: hidden;
+}
+
+p, li {
+  line-height: 1.5;
+}
+
+strong {
+  color: $fg-strong;
+}
+
+blockquote {
+  display: flow-root;
+  font-style: italic;
+  background: $bg-inset;
+  border-radius: $border-radius;
+  padding: 0 $padding-containers;
+  margin-left: 0;
+}
+
+pre, code {
+  font-family: $font-family-code;
+  font-size: $font-size-code;
+  background: $bg-inset;
+  border-radius: $border-radius;
+}
+
+code {
+  padding: 0.2em;
+}
+
+pre {
+  padding: 1em;
+  overflow-x: auto;
+  code {
+    padding: 0;
+    border: none;
+    background: none;
   }
 }
 
@@ -121,6 +187,7 @@ table {
   display: flow-root;
   border-radius: $border-radius;
   padding: 0 $padding-containers;
+  margin-bottom: $padding-containers;
   &.type-info {
     background: $bg-info;
   }
@@ -129,8 +196,60 @@ table {
   }
 }
 
-.breadcrumbs {
+.top-header {
+  display: flex;
+  flex-direction: row;
+  align-items: baseline;
+  padding: $padding-containers-small $padding-containers;
   background: $bg-header;
+  justify-content: space-between;
+  border-bottom: 1px solid $border-color-header;
+
+  .c20-logo {
+    font-family: $font-family-headers;
+    background-image: url("/assets/librarian.png");
+    background-repeat: no-repeat;
+    background-size: contain;
+    padding-left: 48px;
+    margin-right: 1em;
+    line-height: 40px;
+    font-size: $font-size-logo;
+    text-decoration: none;
+    font-weight: bold;
+    color: $fg;
+  }
+
+  .c20-name-short {
+    display: none;
+  }
+  .c20-name-long {
+    display: inline-block;
+  }
+
+  .c20-top-nav {
+    display: flex;
+    flex-direction: row;
+    align-items: baseline;
+
+    a {
+      margin-left: 1em;
+    }
+  }
+}
+
+.content-sidebar {
+  grid-area: sidebar;
+  padding: $padding-containers;
+  max-width: 20em;
+}
+
+.content-main {
+  grid-area: main;
+  display: flex;
+  flex-direction: column;
+}
+
+.breadcrumbs {
   padding: $padding-containers;
   ul {
     list-style: none;
@@ -140,103 +259,92 @@ table {
   }
   li {
     display: inline-block;
+    &:first-child::before {
+      content: none;
+    }
     &::before {
-      content: '/';
-      padding: 0 8px;
+      content: '›';
+      display: inline-block;
+      font-weight: bold;
+      font-size: 1rem;
+      margin: 0 5px;
     }
   }
 }
 
-.metabox, figure, aside {
-  float: right;
-  clear: right;
-  width: 35%;
-  background: $bg-overlay;
-  border: 1px solid $border-color;
-  border-bottom-width: 0;
-  border-radius: $border-radius;
-  margin: 0 0 $padding-containers $padding-containers;
-
-  figcaption {
-    font-style: italic;
-  }
-  section, figcaption {
-    display: flow-root;
-    border-bottom: 1px solid $border-color;
-    padding: 0 $padding-containers;
-  }
-  section.header {
-    text-align: center;
-  }
-  section.img, img {
-    padding: 0;
-  }
-}
-
-.content {
-  padding: 0 $padding-containers $padding-containers $padding-containers;
+.content-article {
+  max-width: $width-content;
+  margin-right: auto;
   display: flow-root;
+  padding: 0 $padding-containers;
   background: $bg-content;
   border: 1px solid $border-color;
   border-radius: $border-radius $border-radius 0 0;
 
-  h1 {
-    border-bottom: 1px solid $border-color;
-    padding-bottom: 10px;
-    overflow: hidden;
-  }
-
-  p {
-    line-height: 1.5;
-  }
-
-  strong {
-    color: $fg-strong;
-  }
-
-  blockquote {
-    display: flow-root;
-    font-style: italic;
-    background: $bg-inset;
+  .metabox, figure, aside {
+    float: right;
+    clear: right;
+    width: 35%;
+    background: $bg-overlay;
+    border: 1px solid $border-color;
+    border-bottom-width: 0;
     border-radius: $border-radius;
-    padding: 0 $padding-containers;
-    margin-left: 0;
-  }
+    margin: 0 0 $padding-containers $padding-containers;
 
-  pre, code {
-    font-family: $font-family-code;
-    font-size: $font-size-code;
-    background: $bg-inset;
-    border-radius: $border-radius;
-  }
-
-  code {
-    padding: 0.2em;
-  }
-
-  pre {
-    padding: 1em;
-    overflow-x: auto;
-    code {
+    figcaption {
+      font-style: italic;
+    }
+    section, figcaption {
+      display: flow-root;
+      border-bottom: 1px solid $border-color;
+      padding: 0 $padding-containers;
+    }
+    section.header {
+      text-align: center;
+    }
+    section.img, img {
       padding: 0;
-      border: none;
-      background: none;
     }
   }
 }
 
 .content-footer {
+  grid-area: footer;
   margin-bottom: $padding-containers;
   padding: 0 $padding-containers;
   display: flow-root;
-  background: $bg-overlay;
-  border: 1px solid $border-color;
-  border-radius: 0 0 $border-radius $border-radius;
-  text-align: right;
+  text-align: center;
+  max-width: $width-content;
   border-top: 0;
   color: $fg-faded;
 
   a {
     color: $fg-faded;
+  }
+}
+
+.content-layout {
+  display: grid;
+  grid-template-rows: max-content 1fr;
+  grid-template-columns: 10% 80%;
+  grid-template-areas:
+    "sidebar main"
+    "sidebar footer";
+}
+
+//mobile view
+@media all and (max-width: 40em) {
+  .content-layout {
+    grid-template-columns: 100%;
+    grid-template-areas: "main" "sidebar" "footer";
+  }
+
+  .top-header {
+    .c20-name-short {
+      display: inline-block;
+    }
+    .c20-name-long {
+      display: none;
+    }
   }
 }
