+++ src/content/readme.md
@@ -1,5 +1,5 @@
 ---
-title: Home
+title: The Reclaimers Library
 template: home
 ---
