+++ src/templates/shared/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  wrapper: require("./wrapper"),
+  renderMarkdown: require("./markdown"),
+  metabox: require("./metabox"),
+  ...require("./bits")
+};
